:recycle: (FE:Id) id を文字列ではなく数値型として扱うように変更
