+++ frontend/pong/js/api/users/convertUserData.js
@@ -2,7 +2,7 @@ export const convertUserData = (userData) => {
   const { id, username, display_name, avatar } = userData;
 
   return {
-    id: id.toString(),
+    id,
     username,
     displayName: display_name,
     avatar,
