+++ frontend/pong/js/api/users/getUser.js
@@ -1,12 +1,13 @@
 import { Endpoints } from "../../constants/Endpoints";
+import { isValidId } from "../../utils/isValidId";
 import { fetchData } from "../utils/fetchData";
 import { convertUserData } from "./convertUserData";
 
 export async function getUser(userId) {
-  if (!isValid(userId)) {
+  if (!isValidId(userId)) {
     return {
       user: null,
-      error: new Error("userId: not a 'string' type or empty"),
+      error: new Error("userId: not a positive integer"),
     };
   }
 
@@ -17,6 +18,3 @@ export async function getUser(userId) {
   const user = error ? null : convertUserData(data);
   return { user, error };
 }
-
-const isValid = (userId) =>
-  typeof userId === "string" && userId !== "";
