+++ frontend/pong/js/constants/GeneralConstants.js
@@ -1,0 +1,7 @@
+const Id = {
+  VALID_REGEX: () => /^[0-9]+$/,
+};
+
+export const GeneralConstants = Object.freeze({
+  Id,
+});
