+++ frontend/pong/js/constants/TournamentConstants.js
@@ -1,7 +1,5 @@
-const Id = {
-  VALID_REGEX: () => /^[0-9]+$/,
-};
+const PLAYER_NUM = 4;
 
 export const TournamentConstants = Object.freeze({
-  Id,
+  PLAYER_NUM,
 });
