+++ frontend/pong/js/utils/isNullOrValidId.js
@@ -1,0 +1,3 @@
+import { isValidId } from "./isValidId";
+
+export const isNullOrValidId = (id) => id === null || isValidId(id);
