+++ frontend/pong/js/utils/isValidId.js
@@ -1,0 +1,2 @@
+export const isValidId = (id) =>
+  typeof id === "number" && Number.isInteger(id) && id > 0;
