--- frontend/pong/js/utils/tournament/isValidTournamentId.js
@@ -1,11 +1,0 @@
-import { TournamentConstants } from "../../constants/TournamentConstants";
-
-export const isValidTournamentId = (str) => {
-  if (!(typeof str === "string")) return false;
-
-  const expectedId = Number.parseInt(str);
-  return (
-    TournamentConstants.Id.VALID_REGEX().test(str) &&
-    !Number.isNaN(expectedId)
-  );
-};
